# Avoid high cost crayon functions