The crayon style functions call `has_color()` on every call, which has a
relatively high cost when it is repeated many times; unless the
crayon.enabled option is set. They also use some internal functions e.g.
`%+%` which are fairly costly if there are many calls.

This avoids these costs by

1. caching the result of `has_color()` for the
   entire session. (It can be recomputed by calling `has_color(forget = TRUE)`)

2. Simply using `paste0()` to combine the open and closing escapes with
   the content. This is not as flexible as the crayon implementation,
   but works for the use cases in pillar.

The following benchmark shows these changes result in an order of
magnitude improvement in execution time for the mtcars dataset.

```r
mtcarst <- tibble::as_tibble(mtcars)
r <- git2r::repository()
git2r::checkout(r, "master")
pkgload::load_all()

bnch <- bench::mark(
  base = as.matrix(format.data.frame(mtcars)),
  old = format(mtcarst),
  check = FALSE, min_iterations = 10, filter_gc = FALSE
)

git2r::checkout(r, "fast-style")
pkgload::load_all()

bnch <- rbind(bnch,
  bench::mark(new = format(mtcarst), filter_gc = FALSE, min_iterations = 10)
)

bnch
summary(bnch, relative = TRUE, filter_gc = FALSE)

bnch
#> # A tibble: 3 x 14
#>   expression      min     mean   median      max `itr/sec` mem_alloc  n_gc n_itr total_time
#>   <chr>      <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl> <bch:byt> <dbl> <int>   <bch:tm>
#> 1 base         1.09ms   1.29ms   1.18ms   8.82ms    773.      9.92KB     4   387   500.39ms
#> 2 old        673.58ms 698.03ms 699.43ms 721.19ms      1.43   14.94MB    53    10      6.98s
#> 3 new         72.98ms  80.54ms  78.56ms  92.24ms     12.4    245.6KB     7    10   805.39ms

summary(bnch, relative = TRUE, filter_gc = FALSE)
#> # A tibble: 3 x 14
#>   expression   min  mean median   max `itr/sec` mem_alloc  n_gc n_itr total_time
#>   <chr>      <dbl> <dbl>  <dbl> <dbl>     <dbl>     <dbl> <dbl> <dbl>      <dbl>
#> 1 base         1     1      1     1      540.         1    1     38.7       1   
#> 2 old        617.  540.   593.   81.7      1       1542.  13.2    1        13.9 
#> 3 new         66.8  62.3   66.6  10.5      8.67      24.8  1.75   1         1.61
```