@jimhester 2018-08-17
@gaborcsardi It would be useful to have you review this as well. We might also want to think about having `crayon::has_color()` cache the result like `crayon::num_colors()` does; or switch to using `crayon::num_colors() > 1` in `crayon:::crayon_template()` conditional.

Also likely some of this overhead could be avoided by making more use of vectorization within pillar and reducing the number of calls to each crayon function, but that would require more significant changes in the pillar codebase.