@krlmlr 2018-08-17
Thanks for looking into it. @gaborcsardi: Would you consider providing accessors to the `"_styles"` attribute (and the `"open"` and `"close"` elements) so that we are less dependent on _crayon_'s internals?