@gaborcsardi 2018-08-17
I think the  `start`  and  `finish`  S3 functions do that:
```r
❯ library(crayon)
❯ start(red)
[1] "\033[31m"
❯ finish(red)
[1] "\033[39m"
```