@gaborcsardi 2018-08-17
@krlmlr good idea, btw!