@jimhester 2018-08-20
Sorry, tracked down why switching to `start.crayon()` caused the tests to fail, it is now fixed.