@krlmlr 2018-12-15
Thanks! I rebased your commits and tweaked a bit:

1. Functions are created in `.onLoad()` to avoid build-time dependencies.

1. Converted the subtle and hint styles to use the same framework.

1. `has_color(forget = TRUE)` is called in `colonnade()`, which is the top-level function called by _tibble_ when formatting/printing. This should ensure that output that goes to `sink()` isn't colored.

2. I implemented support for nested formatting, because this seems necessary for now:
    - types and comments are formatted in gray
    - the footer in a too wide tibble shows types in a comment: `style_grey_0.6("# ", ..., style_grey_0.6("<dbl>"), ...)`

There's still room for optimization, but this is a great start.